added a conditional to check if index exists before removing it

+++ db/migrate/20141104145220_remove_and_change_fields.rb
@@ -2,8 +2,8 @@ class RemoveAndChangeFields < ActiveRecord::Migration
   def change
 
    # Rename Indexes
-   remove_index :student_details, [:transcription_coordinator_id]
-   remove_index :student_details, [:transcription_assistant_id]
+   remove_index :student_details, [:transcription_coordinator_id] if index_exists?(:student_details, :transcription_coordinator_id)
+   remove_index :student_details, [:transcription_assistant_id] if index_exists?(:student_details, :transcription_assistant_id)
    add_index :student_details, [:transcription_assistant_id], name: "ta_id"
    add_index :student_details, [:transcription_coordinator_id], name: "tc_id"
